requirements.txt
vastai
python-dotenv